requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
alembic=1.14.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.6.2.post1=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.11.26=hca03da5_0
certifi=2024.8.30=pypi_0
click=8.1.7=pypi_0
dnspython=2.7.0=pypi_0
email-validator=2.2.0=pypi_0
expat=2.6.3=h313beb8_0
fastapi=0.115.5=pypi_0
fastapi-cli=0.0.5=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.0=pypi_0
idna=3.10=pypi_0
jinja2=3.1.4=pypi_0
jose=1.0.0=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
mako=1.3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
mdurl=0.1.2=pypi_0
ncurses=6.4=h313beb8_0
openssl=3.0.15=h80987f9_0
passlib=1.7.4=pypi_0
pip=24.2=py312hca03da5_0
psycopg2-binary=2.9.10=pypi_0
pydantic=2.10.2=pypi_0
pydantic-core=2.27.1=pypi_0
pygments=2.18.0=pypi_0
pyjwt=2.10.1=pypi_0
python=3.12.7=h99e199e_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.19=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h1a28f6b_0
rich=13.9.4=pypi_0
setuptools=75.1.0=py312hca03da5_0
shellingham=1.5.4=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.36=pypi_0
sqlite=3.45.3=h80987f9_0
sqlmodel=0.0.22=pypi_0
starlette=0.41.3=pypi_0
tk=8.6.14=h6ba3021_0
typer=0.14.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024b=h04d1e81_0
uvicorn=0.32.1=pypi_0
uvloop=0.21.0=pypi_0
watchfiles=1.0.0=pypi_0
websockets=14.1=pypi_0
wheel=0.44.0=py312hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
